BRUTE FORCE ATTACK 1.0 - We Are Golden
 * Fixed path demonstrator animation flickering on end.
 * Fixed some GUI animations playing click on highlight.
 * Fixed assembly circle menu upgrade button not being grayed out, if max upgrades was reached.
 * Fixed turrets not being placeable in the buttom of some maps.
 * Rework of enemy movement code structure, resulting in optimizations.
 * Slightly optimized fire particle, still hella slow.
 * Fixed assembly editor hover context not updating correctly.
 * Broke enemy max heal, practically disabled for now.
 * Fixed issue where sound and music would cut out if no data was present already.
 * Changed enemies health to be contained in a long instead, allowing for much larger health pools.
 * Probably a bunch of other stuff as well, didn't expect this much polishing to be done.

+++ Assets/Source/TobyBoss.cs
@@ -7,7 +7,7 @@ public class TobyBoss : MonoBehaviour {
 	public Sprite[] tobySprites;
 	public Enemy enemyComponent;
 	private int spriteIndex;
-	private int startHealth;
+	private long startHealth;
 
 	void Start () {
 		startHealth = enemyComponent.health;
